requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
faiss-cpu==1.13.2
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.15.0
numpy>=1.26.0,<2.0
nvidia-nccl-cu12==2.29.2
openai==2.16.0
packaging==26.0
pandas>=2.0.0
pillow==12.1.0
protobuf==6.33.4
pyarrow==23.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn>=1.3.0
scipy>=1.11.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
streamlit==1.53.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
xgboost>=2.0.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.20
chromadb>=0.4.0